requirements.txt
pydantic==2.12.5
llama-index
llama-index-llms-huggingface
llama-index-embeddings-huggingface
python-dotenv
pillow>=10.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
requests>=2.31.0

llama-index-readers-file
transformers>=4.56.1
torch==2.6.0

# Packages required by baseline_rag/evaluate.py
datasets
ragas
langchain
langchain-community

# Pins to avoid conflicts with other installed packages
openai>=1.65.2,<2.0.0
ipython==8.12.3
tenacity>=8.1.0,<9.0.0
llama-index-core
huggingface-hub
pandas
